fix: usb: gadget: mvebu-u3d: calculate EP queue interval according to BD struct size
bd_interval was calculated originally according to sizeof(bd).

This was correct for a 32bit machine since size of pointer was 0x4,
deriving 0x10 interval, which is the bd struct size.

In 64bit machine it derived 0x8 * 4 = 0x20, which is twice the
bd struct size.
This miscalculation also exceeded coherent dma allocations,
which are correct, since allocated according to bd struct size.

The correct interval is now calculated with sizeof(struct bd).

Change-Id: I794388f29e8c8ff77917fb48e2d078e3962d021b
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32808
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Star_Automation <star@marvell.com>