phy: add phy_send_command callback
This callback might be useful for complicated phy drivers which require
extra operations, in addition to the existing operations (on/off).

Each new operation requires a new call back, which might be
complicated in some cases. "send command" callback helps to
simplify this, be allowing using one callback for many operations.

The commands type is int, this gives the ability to the users
to define the commands for their use.

Change-Id: Ifebe5ec5ca274041bd1d8d4d290fdd79fd0453e9
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/20005
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>