fix: usb: gadget: u3d: avoid duplicated init during resume
In previous implementation, the mvc2_start will be called both
during PM resume and once receive VBUS on interrupt, which leads
to duplication and affect the suspend to RAM long run stability.
This patch modify the PM complete callback that only enable USB3
device irq inside of it.
JIRA A3700-959.

Change-Id: Ieec784b862b0cdd88650be92d239a70e3694800e
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39168
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>