usb: gadget: u3d: add specific compatible entry for Armada 3700
The SoC of Armada 3700 requires different PHY HW reset
settings, compared to other Marvell SoCs, so it is necessary
to change the common compatible string to a new one,
and then it can be distinguished in the driver.

The patch changed the compatible string from
"marvell,mvebu-u3d" to "marvell,armada3700-u3d", and
updated the binding document accordingly.

Change-Id: I2d3ec4ed8952d1048efcb9dc1a0cc12d204c157f
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40250
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>