ata: libahci: set phy mode to SATA before powering it on
Some phy drivers like "marvell,comphy-cp110" are able to configure phy to
various mode, not only particular one. In such case it is required to set
appropriate mode before powering the phy.

For other phys which are only dedicated for SATA the set_mode handler is
not used therefore this patch shouldn't change anything.

Change-Id: Id2244f3903c138313b42e469e3cbdb1f6819f1da
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>