vfio: xhci: add support for XHCI reset hook
The VFIO reset hook is called every time a platform device is
passed to a guest or removed from a guest.
When the XHCI device is unbound from the host, the host driver
disables the XHCI clocks/phys/regulators so when the device is
passed to the guest it becomes dis-functional.
This initial implementation uses the VFIO reset hook to enable
the XHCI clocks/phys/regulators on behalf of the guest.

Change-Id: I89120ea2cf8357bdf2944edad69ee74ce5ded07f
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>