octeontx2-pf: Attach RVU NIX and NPA block LFs
For a RVU PF to function as a NIC, NPA (for Rx buffers,
Tx descriptors etc) and NIX (for rcv, send and completion queues)
are the minimum resources needed. So request admin function (AF)
to attach one each of NIX and NPA LFs (local functions).
Also get the msix vector offsets for these LFs.

Change-Id: I2bae795047aef3b296b8fb01afe057e996bb2f7f
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>