soc: octeontx2: handle Errata 35432
When work is added to a HWGRP it is first buffered in TOAQ and based on
GETWORK pressure the work is moved along the pipeline. Each TAQ buffers
eleven addwork entries. Each HWGRP hold the TAQ buffer until with it till
it is completely used.
Due to the Errata 35432, SSO doesn't release the partially consumed TAQ
buffer used by HWGRP when HWGRP is reset (SSO_AF_LF_HWGRP_RST).
Workaround is to manually drain the TAQ buffer through addwork-getwork loop
till TAQ buffer is completely consumed.

Change-Id: I8a2f7e20cdd33fd19a8534d26ad0ae1b6c0744a7
Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Reviewed-by: Jerin Jacob <jerinj@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1769
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>