octeontx2-pf: Register and handle CGX link notifications
RVU PF and AF (admin function) shares 64KB of reserved memory
region for communication. This region is shared for
 - Messages sent by PF and responses sent by AF.
 - Notifications sent by AF and ACKs sent by PF.

This patch adds infrastructure to handle notifications sent
by AF and adds handlers to process them.

One of the main usecase of notifications from AF is physical
link changes. So this patch adds registration of PF with AF
to receive CGX link status change notifications and also adds
the handler for that notification.

This patch also adds __weak function symbols for mbox handlers to
enable us to add mbox IDs for mbox handlers which are outside the
kernel.

Change-Id: Ide7be9b0d2049807c517138462cf45fdb8a86118
Signed-off-by: Linu Cherian <lcherian@marvell.com>
Signed-off-by: Aleksey Makarov <amakarov@marvell.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>