octeontx2-af: update SSO HWGRP teardown sequence
Teardown HWGRP LF resources used by the consumer in a sane sequence.
To access the LF registers in PF/VF bar2 from AF space we use the AF
ALIAS space based on the slot id of the PF/VF to which the LF is
attached.
Let SSO teardown continue even when NPA_FETCH is not cleared.

Change-Id: Iad4852b4a960b7ea63cf9be0c09539c4ef19eb2d
Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>