scripts: ci: script to create package with all modules
This patch adds a script creating a "$KERNELRELEASE.tar.gz"
file containing kernel-tree with all .ko modules to be placed
into board/target directory "/lib/modules/" known for Kernel for
automatic module install upon usage.

A module and a kernel-Image should have the same built-in
$KERNELRELEASE name. In near all cases the modules and symb-table
are not changed and could be used by different Images. To achieve
this compliance the patch makes the $KERNELRELEASE name "short" --
without adding git-hash.
The resulting kernel-name is the same as an official release-name
 (for example "4.14.76-devel-19.02.0").
Outputs for different versions have this release-name in their
file-name and all are placed into ./out directory.

In addition, the script copies Image and .config files into the
./out/Image_release-name and ./out/config_release-name.

The script may run with 1 optional string-parameter to be added
the output file-names. The built-in $KERNELRELEASE is NOT changed.

Example of usage and outputs:
 Creates files
   out/config_4.14.76-devel-19.02.0
   out/Image_4.14.76-devel-19.02.0
   out/mod_4.14.76-devel-19.02.0.tar.gz
 Creates (adds) files
   out/config_4.14.76-devel-19.02.0-oct
   out/Image_4.14.76-devel-19.02.0-oct
   out/mod_4.14.76-devel-19.02.0-oct.tar.gz

Change-Id: Iddbb0c825b3b3d4852882ccc6e8c81632f068fe6
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/4802
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Yuval Caduri <Yuval.Caduri@cavium.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>