octeontx2-pf: Fixing redundant message from af to pf.
We are using the same interrupt line for mailbox communications in both
up and down directions. When there are two mailbox messages in the same
direction that has to be handled by different mboxes, in cases where
the processing of first message is mid way and interrupt handler gets
invoked on a second message, another work is queued for the first message
again as the num_msgs is not set to zero yet.
Fixing this by introducing another variable that would get updated
inside interrupt handler and setting num_msgs to zero inside interrupt
handler itself.

Change-Id: Ia3a404b312369f8f0f6252236a6cb76ac9364e4c
Signed-off-by: Christina Jacob <cjacob@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/6334
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>