octeontx2-af: Add mbox to read mcam entry
This patch introduces a new mailbox to read a given mcam
entry.This is needed when an mcam entry owner wants to
read-modify-write any one of it's existing mcam entries.

With this patch octeontx2-af driver's consumers need not
store the complete mcam entry configurations locally and
when there is an update in inclusive mcam rule, all it's
overlay mcam rules can be easily updated.

For instance, when af driver's consumer has a request for
MAC address update, along with that it can easily update
other MAC address dependent flow based rules.

Change-Id: I56dad453f0243497a6a62942ca0e11886dcd8e82
Signed-off-by: Vamsi Attunuru <vattunuru@marvell.com>
Signed-off-by: Vivek Sharma <viveksharma@marvell.com>