octeontx2-pf: Fix redundant message from AF to PF
We are using the same interrupt line for mailbox communications in both
up and down directions. When there are two mailbox messages in the same
direction that has to be handled by different mboxes, in cases where
the processing of first message is mid way and interrupt handler gets
invoked on a second message, another work is queued for the first message
again as the num_msgs is not set to zero yet.
Fixing this by introducing another variable that would get updated
inside interrupt handler and setting num_msgs to zero inside interrupt
handler itself.

Change-Id: Iddea099ef872c32946b8bef9ebee860ab8e7713f
Signed-off-by: Christina Jacob <cjacob@marvell.com>