net: mvpp2: move the section which reads hw version after clock init
Access to hw register should be done after all clocks are enabled.
Before this patch in normal circumstances everything works well but in some
it lead to race condition resulting with rcu stall.

The gated clock used for mvpp2 is also shared with axim and because the
axim is treat as AMBA device during it initialization the clock can be
enabled and then disabled. Disabling of mentioned clock happens e.g. after
5s - deferred device timeout in amba bus.

Due to that when the boot time increased (e.g. due to nand bad block
management which can take several seconds) the amba deferred retry func
was scheduled which disabled the clk (the clk was registered only for
AMBA since mvpp2 probe wasn't performed yet).

Performing clock init before any register access in mvpp2 solves the issue
since the second clock consumer is registered (increasing clk count) and
amba driver will not be able to really disable it.

Change-Id: I0ab80d6c2d20750d7a4c1a2343e31c7a1528ce3d
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/7866
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Marcin Wojtas <marcin@marvell.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>