hwmon: scmi: Scale values to target desired HWMON units
If the SCMI firmware implementation is reporting values in a scale that
is different from the HWMON units, we need to scale up or down the value
according to how far apart they are.

Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>
[sudeep.holla: added check of scale = 0 for early exit in scmi_hwmon_scale]
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>

Update:
SCMI sesors headers and sensors.c has been modified with
additional miscellaneous code required by scale conversion functionality.

Signed-off-by: Wojciech Bartczak <wbartczak@marvell.com>
Change-Id: Iacab6b9e2a5f4a21ff29c2013c1c41ab2a918bf9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/65473
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>