net: mvpp2: introduce per-CPU long BM pools
Under certain conditions, with RSS and recycling mechanism
enabled and some of the traffic terminated / migrating
between the CPUs, a memory leak can be observed.

The debug showed a most likely root cause to be
ordering-dependency between packets passed over
L3/4 stack and free_frag executed on same/different CPUs.
Any unbalancing may cause page cache fragmentation which
not always could be reclaimed by kswapd or reclaimed very late.

In order to prevent above phenomenon, introduce per-CPU
Buffer Management (BM) pools for affected packet size, i.e.
the "long" pools. For that purpose maintain such pools
separately and map to the RXQs, depending on whether
jumbo frames are used or not.

By default this mechanism will remain disabled as
it collides with MUSDK BM pool mapping. In order to
use it, 'mvpp2.recycle=1' module param enabler
should be added to the command line. It will however
disable changing to the MUSDK mode and the user will
be informed with WARN() when attempting to do it.

Based on debug of Yan Markman <yanm@marvell.com>

Change-Id: I52ac4a4222f47d63383d6ede811fba4ae37f698e
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/8846
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>