octeontx2-af: Disable MCAM entries during NIX_LF_FREE
Ntuple filters installed by user for a interface needs
to be deleted by the user but currently those are
deleted when interface is brought down say using
ifconfig ethx down. Correct behavior is entries
which target the down interface should be disabled
hence traffic to that interface is dropped at MCAM.
To fix the problem:
 - A flag is added in NIXLF_FREE message for only disabling
   MCAM entries.
 - MCAM entries of a PF/VF are deleted during its FLR.
 - For a VFn, entries which are owned by its parent PF
   but direct traffic to VFn are disabled during FLR.
 - Disabled entries are enabled when NIXLF_START mbox
   message is called for the interface.

Also some trivial changes like renaming variables are done.

With these changes VF interface need not be initialized when
installing a ntuple filter by its PF.

Also debugfs is updated to show enable/disable status and
target of an entry. MCAM rules are now displayed as:
        PF1
        target: PF1 VF0
        mcam entry: 227
        ether type  0x8100 0xffff
        action: Direct to queue 0
        enabled: yes
        counter: 1
        hits: 0

Change-Id: I5f7716c9b5b9e3899b9bdbcf727df1fdbeea5324
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/10406
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>