octeontx: increment MBOX major version
The MBOX did not change, however, the semantics did.
Since the decision to keep all devices open in ODP and perform
domain reset on any device closure in kernel we want to ensure that
someone does not run older ODP, which may close devices on a whim,
with newer kernel, which performs domain reset more often.

Change-Id: I723ae0c303e8a30782eff6d951699d96b049e80e
Signed-off-by: Angela Czubak <aczubak@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/10194
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Reviewed-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>