octeontx2-af: Transmit packets during SMQ flush
NIX SMQ flush may not complete if no send queue is
transmitting packets. Hence packets need to be
transmitted from a send queue when SMQ flush is
initiated. This patch adds a SMQVF driver which
is a stripped down version of VF netdev driver.
The driver exports a function which can be called
by AF driver during SMQ flush to transmit packets.
A new VF of AF with pcifunc 17 is used for this
purpose. So 17th AF VF should not attached to
userspace.

Change-Id: I6592f7615f33f423de2d464f7441864443f0dec4
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/14560
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>