PCI: aardvark: Implement work-around for PCIe completion timeout errata
The Armada 3700 Errata document describes an errata numbered 3.12 and
called "PCIe completion timeout". The errata says:

""
The PCIe IP supports a strong-ordered model for inbound posted vs.
outbound completions.

Due to this erratum, this feature is not supported. With the strong
ordering rule, completion cannot pass posted write.
""

This errata has been encountered by users of the Armada 3720 platform
with use-cases where the PCIe interface is used to connect WiFi chips.
Under significant load, incorrect behavior was encountered, and traced
back to this errata.

The recommended work-around from the AArdvark designers is to set bit
DIS_ORD_CHECK in the "Debug and Mux Control Register",
which "disables ordering checks".

Change-Id: I4117e13e97de40731a7118e899d2dc872d3bafea
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Reviewed-on: https://sj1git1.cavium.com/15560
Reviewed-by: Grzegorz Jaszczyk <jaszczyk@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>