scmi: driver: insufficient buffer size for rx
scmi cpu-freq driver allows MAX_OPPS (16) OPPs but when transferring
the 16 OPP data from SCP to ARM, because of the max_msg_size limited
to 128, it fails to convey the information about all 16 OPPs.

An increase to 256 for max_msg_size accommodates all entries correctly
and linux receives them with integrity.

Bug: CPCSCPMCP-20

Change-Id: Ie7f0554c91948a52473ee9143dc068106fd5ef30
Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/16842
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>