arm64: Add workaround for Marvell erratum 37119
Enable workaround for erratum 37119.  On some OcteonTX2 chips,
while using the SSO workslots in userspace can cause problems
as the cache are not invalidated correctly.  To workaround
this erratum, a local tlbi is placed before the eret.
This TLBI does not need to invalidate anything so a tlbi against
asid 0 is used.

Modified the midr range helpers according to upstream
commit 41b3073644e3
("arm64: capabilities: Clean up midr range helpers")

Change-Id: I0313daae25e8ccaaffb84f9b6db5e6859b693dfa
Signed-off-by: Andrew Pinski <apinski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17085
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>