octeontx2-af: kpu profile add support for HIGIG2 protocol
Add support for HIGIG2 protocol header to the kpu parser profile

Add a NPC_LT_LA_HIGIG2_ETHER for a combined header of HIGIG2 and
Ethernet.  Add a NPC_LT_LA_IH_NIX_HIGIG2_ETHER for a combined
header of nix_ih, HIGIG2 and Ethernet.

Also add 2 upper flags in LA to indicate the presence of nix_ih
and HIGIG2.

Change-Id: I097688128e238c5f61d60b99eecc8967c47b05db
Signed-off-by: Hao Zheng <haoz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17840
Reviewed-by: Kiran Kumar Kokkilagadda <Kiran.Kokkilagadda@cavium.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>