mailbox: Enable interrupt handling
SCP is AVS bus master regulating the VRM. In case of an error
detected on AVS bus during the periodic communication from SCP
to VRM, erros are anticipated. SCP now raises an interrupt to
AP in case of any such error detected.

The associated ISR in thread context at this moment just prints
a kernel error message but any corrective action if needed by AP
could be added in the threaded ISR function as this is just the
place holder.

Bug: CPCSCPMCP-34

Change-Id: I1ccdd24e940fd359309d8362bd0d1b4966ae91c6
Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17575
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>