octeontx2-pf: Fix wrong info in ethtool's list of supported link modes
In otx2_get_link_mode_info(), the signed integer literal "1" is left-
shifted by a value obtained by the table lookup
"cgx_link_mode[bit_position]".  The result is then implicitly promoted from
signed int to u64 before it gets bitwise-ORed and assigned to
"ethtool_link_mode" (whose type is u64).

This is problematic if the result of the table lookup (which is used as the
number of bit positions to left shift) is greater than or equal to 31 (one
less than the bitwidth of an int).  Such a situation will lead to a wrong
value for "ethtool_link_mode", causing ethtool to show wrong or incomplete
information in the list of supported link modes.

Fix it by adding the "ULL" suffix to "1".

Change-Id: I411befd4ad23df2566bf5e9cd6141faeeb53a26c
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17524
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
(cherry picked from commit 8b1d4774140e9b1ff99a49e2b6ab8b610631c2e9)
Reviewed-on: https://sj1git1.cavium.com/17537
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>