octeontx2-rm: remove CPT LF resource management
CPT LFs will be assigned to VFs of CPT PF instead
of VFs of RM PF.
The requirement for this change is:
multiple cpt mbox messages are handled in the CPT PF driver.
to use these message, CPT LF should be linked to VFs of CPT PF only.

domain will receive the DBDF of CPT VF as a param.
RM will treat the DBDF of CPT VF in a way similar to DBDF of ports.

Change-Id: Ia2f2e86d1d9784276cad8f77679512f40956c02d
Signed-off-by: Vidya Velumuri <vidya.velumuri@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17735
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Reviewed-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>