net: mvpp2: cleanup possible residues in the BM left by the firmware
It may happen that the OS is booted without BM pools cleanup,
e.g. via old U-Boot or UEFI. It turned out that it is also a
problem when booting the kernel with U-Boot's bootefi.
Avoid the issue by making sure the BM pools are completely empty
and disabled prior to using in the mvpp2 driver.

Change-Id: Ie1389d812b7761d3abb250fa7dfd6ba8a7e6b01d
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/18185
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>