net: mvpp2: update and suppress RX error debug
According to the documentation the 0x0 error code
of RX descriptors means 'MAC error'. Hitherto designation
('CRC') is only one type of it, so update the naming
to be more generic.

This patch changes also the default behavior - by default
the counters are updated upon errors, but the print is
now available only with the debugs enabled.

Change-Id: I5e1fc3fb50b705a4b035f9bff60529be63c27801
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/19004
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>