octeontx2-pf: Do not set num_msgs in mbox header
A single interrupt line from VF to PF is used by
VF to send messages to PF and for replying to UP messages
sent by PF. Hence num_msgs in mbox header is cleared
explicitly in PF interrupt handler to avoid queuing a
work twice for an event. The same num_msgs which was
cleared in interrupt handler was set again in work
queue handler by mistake. This patch fixes it.
Also
 - clearing interrupt in work queue handlers is not
   necessary hence removed it.
 - checking status in message response is not needed
   hence removed it.

Change-Id: I1875d37816b61b4874eea1c7175c63dfea71bcb1
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/20142
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>