octeontx2-af: Get block address from slot number
Currently there is no way in AF driver to get
the RVU block address of a attached LF if
caller has LFs from two blocks of same type.
Hence use the slot number of LF from the
caller to retrieve the block address of block
from which the LF is attached. It is the
responsibility of caller to provide a global
slot number in mailbox.
Say a PF has 3 LFs from CPT0 and 2 LFs from
CPT1. Then slot number from caller is used as
below:
	Slot 0..2 -> block address CPT0
	Slot 3..4 -> block address CPT1

Change-Id: Ia88d7d2df11c5a142cd1289429a1d57a1ea957ba
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/27988
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>