drivers/uio: fix the MAX_BAR_MAP value
The MAX_BAR_MAP should reflect the total number of PCIe BARs,
which is 6. The first 2 BARs are using 64-bit addresses, so
they occupy 2 BAR slots each, resulting in total BARs
number of 4.
Limiting BAR ID to 4 results in maximum 2 64-bit BAR entries,
which is unnecessary limiting the driver capabilities

Change-Id: I93a978e3c600194095994052a73e56e4b7f48159
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/21465
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>