octeontx2-af: Fix rvu_reg_poll() to poll reg atleast twice
Currently if on the first check the operation is still not
finished, the poll goes to sleep for 2-5 usecs. But if for
some reason (due to secure interrupts etc) by the time the poll
wakes up the 10ms time is expired then we don't check if operation
is finished or not and return failure.

This patch fixes this to check HW operation after sleep so that
the status is checked atleast twice.

Change-Id: Ie05e21536b6a80738d4473617dad88e20f0b84eb
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/22474
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>