net: thunderx: Fix transmit channel config for secondary qsets
Due to the changes introduced for loopback (LBK) interface
support, VNIC's secondary qsets transmit path is broken. This
is due to the incorrect way of checking LBK mapped VF which
resulted in configuring transmit channel as PKO instead of
CGX for secondary Qset SQs.

This patch fixes the issue, also cleaned up the LBK VF checks
throughout the PF driver.

Change-Id: Ica128211315287137ae82cee7346a9e018a09df0
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/23679
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Harman Kalra <Harman.Kalra@cavium.com>