net: thunderx: Do a PCS reset upon SGMII link toggle
When physical link toggles with BGX mode being SGMII, BGX to Serdes link
needs to be reset for it function normally otherwise packet IO won't work.

This patch does the PCS reset to get IO functioning again When the traffic
is running, if we unplug the Ethernet cable from SGMII port and plug it
again, it become malfunction.

Change-Id: Ic5eaddd3b583f43e9568a6da532e638f4eb61116
Signed-off-by: Prakash Brahmajyosyula <bprakash@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/10407
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>