drivers: marvell: otx2-sdei-ghes: Set GHES block_status correctly.
The GHES block_status is only required to contain the count
of entries contained in the data section.
The error status bits are optional and only server to enhance
the Error Severity of each individual entry.
Therefore, this commit removes the hard-coded
'ACPI_HEST_CORRECTABLE' status as it is neither correct nor
required.

Change-Id: I0fc15d1c4b879333fd9cb3546d4f917ddf4a5450
Signed-off-by: Rick Farrington <Ricardo.Farrington@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/26638
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>