octeontx: Insert nops and a load after stores to SWTAG
An issue exists for 83xx pass 1.1. and 1.2 where a killed instruction
that is a main micro-TLB miss may also update the switch-tag micro
TLB. A subsequent load may then be misinterpreted as a load to
SSOW_VHWS()_SWTP, and return incorrect data.
Insert 13 nops, followed by a load, to clean up the micro-TLB state.

Signed-off-by: Angela Czubak <aczubak@marvell.com>
Change-Id: I473f80b24479935b8fd344a9e3fe177b98885bb9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/27340
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>