octeontx2-af: Add new Mbox messages and handlers for new REE blocks
Mailbox support for REE0 and REE1 blocks in 98xx platform.
This patch adds new REE mailbox messages to mbox.h:
- Config LF
- REE BAR0 registers read and write
- REE get Rule database length
- REE get Rule database for incremental programming
- REE Rule database programming

This patch adds the corresponding new mbox handlers to rvu_ree.c

The rule database that is passed via mbox contains instruction of
address and data and is programmed to REE.
For programming REE allocates memory that is accessed by HW:
- 128-MB structure containing the complete compiled regular-expression
  graphs
- AF Queue that points to 16KB blocks that hold filtering instructions

REE also allocates memory for keeping the rule database that was
received via mbox. It is used for immediate programming and future
incremental programming

Change-Id: I3679b736bafac8ba944f47252dad52fb2fce6506
Signed-off-by: Smadar Fuks <smadarf@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/29301
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>