drivers: gicv3: Adds workaround for Marvell erratum 38545
On some OcteonTX2 chips, an issue exists wherein when an interrupt
is pending and a second interrupt arrives at the core under certain
stall conditions, a read of interrpt acknowledge register will return
the first interrupt but not acknowledge it. A second read of IAR will
return the first interrupt again, not the second interrupt.

Updated the midr range helpers according to upstream
commit 41b3073644e3
("arm64: capabilities: Clean up midr range helpers")

Change-Id: I8929a93ffbff41773161ac2b5c2b77863259c61b
Signed-off-by: Bhaskara Budiredla <bbudiredla@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/31197
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Linu Cherian <lcherian@marvell.com>
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>