coresight: tmc: add check for secure ETR access
The current TMC-ETR driver checks only for non-secure invasive debug
capabilities and if these bits are not set in AUTHSTATUS register it
returns -EACCESS.

Since OcteonTx2 supports only secure invasive debug this causes the
driver to return -EACCESS and not proceed with further configuration.

Fix this issue by adding a new function that checks for secure invasive
debug capabilities in the AUTHSTATUS register and return -EACCESS only
if none of the secure/non-secure capabilities are supported.

Change-Id: I66926d9890204261fa0f3e2668e19bfd6ca1e6cb
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/32255
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Linu Cherian <lcherian@marvell.com>