iommu/arm-smmu: Workaround for Marvell Armada-AP80x SoC erratum #582743
Due to erratum #582743, the Marvell Armada-AP80x can't access 64bit to
ARM SMMUv2 registers.

This patch replaces the writeq/readq with hi_lo_* functions for
MRVL_MMU500 model.

We also mask the MMU_IDR2.PTFSv8 fields to not use AArch64 format but
only AARCH32_L. Indeed with AArch64 format 32 bits access is not
supported.

Note that separate writes/reads to 2 is not problem regards to
atomicity, because the driver use the readq/writeq while initialize
the SMMU, report for SMMU fault, and use spinlock in one
case (iova_to_phys).

Change-Id: I07209c5adcf1d73a088082193f4b15a863db2104
Signed-off-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/34951
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>