cn10k-cpt: Workaround MBOX timeout
CPT PF sends bunch of mailbox messages(~64) to AF at a time
to read/write the CPT engine registers and waits for mailbox
response for all the messages. On cn10k asim, some times
mailbox response time is crossing the MBOX_RSP_TIMEOUT value.
So, modified code to send the mailbox message one at a time
as a workaround.

Signed-off-by: Srujana Challa <schalla@marvell.com>
Change-Id: I451dbcd90f0856feb4219264926b12d9604bb975
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/36933
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>