octeontx2-npa: Fix NPA PF driver commit.
A header file advertising the API exported by NPA PF driver introduced
in the commit "8465a16348e9", is missed in that commit. This made the
driver unusable by components that need this driver. Added missing file
in this commit and h/w units that need pool manager can simply include
this file and use the API.

Fixes: 8465a16348e9("soc: octeontx2: Add NPA PF driver")
Signed-off-by: Nalla, Pradeep <pnalla@marvell.com>
Change-Id: I524376bc31ee0480fd1428a91d1ed079f29377e9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/37686
Reviewed-by: Radha Chintakuntla <radhac@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>