x86/resctrl: Track the closid with the rmid
RDT's rmid are independent of the closid, an rmid can be allocated,
used and freed without considering the closid.

MPAM's equivalent feature is pmg, which is not an independent number,
it extends the closid space. For MPAM, only pmg-bits worth of rmid
can be allocated for a single closid. e.g. if there is 1 bit of pmg
space, then each closid can have two rmid.

To allow rescrl to disambiguate rmid values for different closid,
everything in resctrl that keeps an rmid value needs to know the closid
too. This will always be ignored on x86.

Signed-off-by: James Morse <james.morse@arm.com>