octeontx2-af: don't enable MCAM entries when target is not initialized
This patch checks whether the target nixlf of a rule is initialized or
not and based on that the entry is enabled or disabled regardless of
requester's enable bit. This fixes the SMMU translation faults observed
when PF installs flow steering rule to forward traffic to VF which either
doesn't have a nixlf attached or nixlf's queues are not initialized.

Change-Id: Id5dfae1427c40452002a845dff21efe0b939a1a2
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/42301
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>