octeontx2-af: don't enable MCAM entries when target is not initialized
This patch checks whether the target nixlf of a rule is initialized or
not and based on that the entry is enabled or disabled regardless of
requester's enable bit. This fixes the SMMU translation faults observed
when PF installs flow steering rule to forward traffic to VF which either
doesn't have a nixlf attached or nixlf's queues are not initialized.

Change-Id: I61b236aa6aaf6aa00682bfe7adc537d8d05a4305
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/42340
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>