octeontx2-af: Introduce internal packet switching
As of now any communication between CGXs PFs and
their VFs within the system is possible only by
external switches sending packets back to the
system. This patch adds internal switching support.
Broadcast pkt replication is not covered here.
RVU admin function (AF) maintains MAC addresses
of all interfaces in the system. When switching is
enabled, MCAM entries are allocated to install rules
such that packets with DMAC matching any of the
internal interface MAC addresses is punted back
into the system via the loopback channel.
On the receive side the default unicast rules
are modified to not check for ingress channel.
So any packet with matching DMAC irrespective of
which interface it is coming from will be forwarded
to the respective PF/VF interface.
The transmit side rules and default unicast rules
are updated if user changes MAC address of an interface.

This feature is not enabled by default to save MCAM
entries. Since transmit side rules need to be at the
bottom of the MCAM, it is recommended to enable this
feature before any of the PF/VF interfaces are
initialized. Otherwise the switching rules might
impact other MCAM entries installed.

To enable switching use below command:
devlink dev eswitch set pci/0002:01:00.0 mode switchdev
To disable:
devlink dev eswitch set pci/0002:01:00.0 mode legacy

Change-Id: I3bd28dc92c86e1296eb3d1edfa8600d64907c1be
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/47679
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>