spi: Fix possible issue in xSPI driver
When xSPI will be configured by probe, and before any operation
on spi memory SMC call will be executed - it will cause an error, as
xSPI driver will not check internal state of IP core before issuing
STIG command.
Fix issue - verify if xSPI is in STIG mode, and run config again if
it is not.

Change-Id: I03a8eab23f0f67b977109453c47f1532c3175321
Signed-off-by: Witold Sadowski <wsadowski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/47937
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/48123
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>