arm64: dts: cn913x: add device trees for topology B boards
The CN913x DB with topology B is similar to a regular setup (A)
boards, but uses NAND flash as a boot device, while topology A
boards are booting from SPI flash.
Since NAND and SPI on CN913x DB boards share some wires, they
cannot be activated simultaneously.
The DTS files for setup "B" are based on setup "A", in which the
CP0 NAND controller enabled and CP0 SPI1 disabled.

Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Change-Id: I7db4ad0da28bf5798162ba0ea5c7af7d61cedcc1
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/48231
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>