net: phy: 10G mode for fixed PHYs
The port 9 & 10 on a marvell DSA switch needs to be connected in
fixed-link mode to get 10G speed enabled. This mode uses an emumated
PHY that exposes standard registers. However, multiGigabit PHYs tend
to use standardised C45 registers to exposes these modes, which aren't
supported in swphy.

We could add these modes, and that will probably be done in mainline,
but in a way that is very difficult to backport. It implies reworking
the way we represent link-modes, which impacts all drivers in
net/ethernet and net/phy.

For now, add a way to simulate this feature by hardcoding the
phy_interface_modes that would need such PHYs to be emulated.

Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Change-Id: Ib2beb9a114c69783c93ce90ddc466853df75887e
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/50053
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <SunilKovvuri.Goutham@cavium.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>