octeontx2-af: Fix ending channel number calculation for SDP VF
There is a logic to verify if the channel falls in the SDP channel
range. A minor mistake in calculating the ending channel was causing
failure for a case where we use all the SDP channels and last VF tries
to allocate the last channel. This patch fixes the mistake.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: Iff251fe2abcf305b7df60183fe541389e436d2bd
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/52671
Reviewed-by: Satananda Burla <sburla@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>