octeontx2-pf: Implement ntuple filter for any non-zero ethertype
In scenarios like if a packet arrived on a link do not
match the criteria of any of the installed ntuple filters
then that packet must be steered to a PF/VF of that link.
To implement this case allow installing filters with
non-zero L2 ethertype field and dont care mask so
that packets with any ethertype will match the filter.
This way all packets can be steered to a PF/VF by
default since every packet has a ethertype.

Usage:
ethtool -U eth0 flow-type ether proto 0xffff m 0xffff \
vf 0 queue 0

Note that above ntuple filter should always be at lower
priority compared to rest of the installed filters.

Change-Id: I636fd386eadf498266090275efe82d91c9bb3015
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/59326
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>