x86/resctrl: Allow resctrl_arch_rmid_read() to sleep
MPAM'sc cache occupancy counters can take a little while to settle once
the monitor has been configured. The maximum settling time is described
to the driver via a firmware table.

To avoid exposing this to resctrl, it should be hidden behind MPAM's
resctrl_arch_rmid_read(). But add_rmid_to_limbo() calls
resctrl_arch_rmid_read() from a non-preemptible context.
add_rmid_to_limbo() is opportunistically reading the L3 occupancy counter
on this domain to avoid adding the rmid to limbo if this domain's value
has drifted below resctrl_rmid_realloc_threshold since the limbo handler
last ran. 'This domain' involves disabling preeption to prevent the
thread being migrated to CPUs in a different domain between the check
and resctrl_arch_rmid_read() call.

Instead, unconditionally call resctrl_arch_rmid_read(), and get it to
read the arch specific counter via IPI if its called on a CPU outside the
domain. This change results in more work for platforms with multiple
domains, but ensures an rmid is never added to the limbo list when it
is clean for a remote domain.

Signed-off-by: James Morse <james.morse@arm.com>