octeontx2-pf: Use CQ_OP_STATUS register for CQE count
Current driver uses software cq head pointer to polls on CQE header
in memory to determine if CQE is valid. Software on Arm needs to make
sure, that the reads of the CQE do not get re-ordered so much that it
ends up with an inconsistent view of the CQE. To ensure that DMB barrier
after read to first CQE cacheline and before reading of the rest of the CQE
is needed. But this will impact the performance, instead use hardware
CQ head and tail pointers to find the valid number of CQEs.

Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: I95df0551bdc5feaffbcbbc870b2c6ed9e029e42d
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/61487
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>