soc: marvell: cn10k SerDes: ioctl interface for serdes-cli
serdes-cli application must send notification to
ECP when entering/leaving prbs mode or changing the
loopback state. It is done through SMC call to
ATF. Since SMC calls can only be triggered from
EL1 or higher privilege levels, it needs to be
delegated to the associated kernel driver. Hence
the changes in this patch to support sending ECP
notifications via SMC call on behalf of the
application which will communicate over ioctl
interface with the driver.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: I06b68bd58136ffbe2fe613cd12fcee552b27f394
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/66332
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>