octeontx2-bphy-netdev: ptp: Fix ptp timestamp reading across lmacs
The RX GPINT is common for all LMAC's per RFOE MHAB. So the RX packet
handler uses the netdev_priv structure of first valid LMAC until the
point where the code reads the PSW and get's the actual lmacid. From
that point onwards, the code uses the netdev_priv structure of that
lmacid. There is an issue where the code checks whether PTP is enabled
or not for that lmacid before learning the actual lmacid on which the
packet is received. As a result, the PTP timestamp will be always passed
as 0 to the ptp4l application. This patch fixes the issue by reading
psw timestamp unconditionally for all packets and send the timestamp
to ptp4l application properly.

Change-Id: I6790dac2ea09ac5ab3caf401712f7728f911e986
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/69270
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>