octeontx2-pf: Add support for HTB offload
This patch adds support for HTB offload callbacks. There are
total 16 transmit queues reserved for HTB offload. The priority
and quantum are also supported. The hardware supports strict
priority among the TLx queues. The priority values 0 to 6 can be
used to configure priority among the leaf nodes. The priority value
7 is used for DWRR. There is a hardware limitation where only one
of the priority can be used for DWRR. The TC_HTB_NODE_MODIFY is not
implemened.

Change-Id: Id27c36df545273cc1af5d23860abbcdf6d408a1c
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/71878
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>