octeontx2-af: use clock source to compute start cycle
User land needs exact clock cycle at which TIM admin function acknowledged
the TIM ring enable from 0 to 1. This can be calculated by using the next
expiry offset that is loaded into TIM_AF_RINGX_CTL0 when enable transitions
from 0 to 1.
Since, expiry offset is 32bit and clock source is 64bit, capture the upper
half of the clock counter before and after the enable transition and retry
if we are out of sync.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Change-Id: I781b8bfb1e265119d75c199b3c2fcedf728d2410
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/72845
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>