PCI: controller: Add OcteonTx2 PCIe Endpoint driver
This patch adds a driver that aids in configuring the PEM when OcteonTx2
is in PCIe Endpoint mode. Most of the endpoint configuration is done in
firmware. The driver sets up the BAR4 space which is 4MB chunks in 16
indices. This region will be used as mailbox kind of communication
channel from host to endpoint.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: I091419d9c791133d3ebfc67c81a9e6026359675c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/73208
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>