PCI: octeontx-83: configure PKI sriov
Octeontx 83xx has a unique case, where PKI does DMA with more than 1
streamids, but PKI is not a SRIOV/ multi function device.
this feet is acheived by giving 8 bit programmable streamid in QPG tables.
The intention was to allow differnt addressspace per GMID.
but in reality no one accepts a device doing DMA with random streamids.
and there is no infrsastrutre in Linux to allow this.

This patch address this problem by emulating SRIOV on PKI.
With this patch in place, applications will bind one of the PKI VF
to allow DMA.

Change-Id: If52ab09c9bbb08e16750fbdd0cf0f539d616cd56
Signed-off-by: Harman Kalra <hkalra@marvell.com>