net: thunderx: Enable independent Rx packet size limit setting per LMAC
As of now MTU is enforced only on the transmit side, but on Rx
path packets of size 'NIC_HW_MAX_FRS' are allowed. This patch
changes the behavior to restrict Rx packets also to MTU size.
For this a separate PKIND is assigned to each LMAC and max length
in PKIND is set to MTU.

Change-Id: I852a8171b6fe0e4d787e5021c74fec91ed7b6482
Signed-off-by: Sunil Goutham <sgoutham@cavium.com>
Signed-off-by: Maciej Czekaj <maciej.czekaj@cavium.com>
Signed-off-by: Harman Kalra <hkalra@marvell.com>