octeontx2-af: Fix NIX link credits
When number of LMACs active on a CGX/RPM are 3, then
current NIX link credit config based on per lmac fifo
length which inturn  is calculated as
'lmac_fifo_len = total_fifo_len / 3', is incorrect. In HW
one of the LMAC gets half of the FIFO and rest gets 1/4th.

Change-Id: I724679ddcfa506e0a44b1c016d452c83d9808cb9
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/78542
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>