octeontx2-pf: Allow htb operations irrespective of interface state
Currently the HTB commands are not accepted when the interface is in
DOWN state. This patch adds support for allowing HTB commands
irrespective of interface state. When the interface is DOWN, the driver
will allocate the HW resources but will not enable the tree. When the
interface is brought UP from DOWN state, the QoS txschq tree is enabled
to send traffic.

Change-Id: I7d20337a1bfc567a4d50d51f0fc5101a9bbe4af3
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/78752
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>