mailbox: marvell: Initialize interrupts only if there's client for data
SCP interrupts should be enabled only when the client (AP) can receive
the data. This happens when the MHU instance and mailbox channel
are established and initialized correctly.
To ensure this enabling SCP interrupts has been postponed.
Interrupts are enabled in startup callback of mailbox channel.

Without this change there is a race between first SCP interrupt and
mailbox initialization code. In some cases this leads to crash.
Mostly when interrupts comes before initialization completes.

Signed-off-by: Wojciech Bartczak <wbartczak@marvell.com>
Change-Id: Iefcc37edc004a4e76b0383488529332f314d5435
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/81119
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>