arm64: dts: imx8dxl-v2x: enable gps
MIA-M10Q gnss receiver is connected to both uart2 and i2c2.
Describe both as sub-nodes, with the i2c node commented - as it does not
have a driver.
The uart child shows up as a /dev/gnss0 device.

For now the enable-gpios is skipped because the currently wired gpio is
not usable as an output, and will be changed in the future.

Signed-off-by: Josua Mayer <josua@solid-run.com>